specific typing

+++ 2015/src/common/testing/testing.c
@@ -1,16 +1,17 @@
 #include "testing.h"
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
-int eval_test(const char* re_name, const char* te_name) {
+uint8_t eval_test(const char* re_name, const char* te_name) {
     const registry_entry* re_curr = find_re(re_name);
     const test_entry*     te_curr = find_te(re_curr, te_name);
 
-    int r = re_curr->test_run(te_curr);
+    int64_t r = re_curr->test_run(te_curr);
     if (r != te_curr->expected) {
-        fprintf(stderr, "'%s' failed: expected %d, got %d\n", te_name, te_curr->expected, r);
+        fprintf(stderr, "'%s' failed: expected %lld, got %lld\n", te_name, te_curr->expected, r);
         return 1;
     }
     return 0;
