+++ 2015/src/day01/day01.h
@@ -1,5 +1,7 @@
 #pragma once
 
-void day01_main(void);
-int  day01_part1(const char*);
-int  day01_part2(const char*);
+#include <stdint.h>
+
+void    day01_main(void);
+int16_t day01_part1(const char*);
+int16_t day01_part2(const char*);
