+++ 2015/src/day02/day02_test.c
@@ -2,6 +2,8 @@
 #include "testing.h"
 #include "utils.h"
 
+#include <stdint.h>
+
 const test_entry* day02_tests[] = {
     TE("./data/example01.txt", "d02_p1_e01", one, 58),
     TE("./data/example02.txt", "d02_p1_e02", one, 43),
@@ -11,9 +13,9 @@ const test_entry* day02_tests[] = {
     TE("./data/input.txt", "d02_p2_i", two, 3783758),
 };
 
-int day02_test_run(const test_entry* te) {
-    char** input = file_to_array(te->file);
-    int    r     = (te->p == one) ? day02_part1(input) : day02_part2(input);
+int64_t day02_test_run(const test_entry* te) {
+    char**  input = file_to_array(te->file);
+    int64_t r     = (te->p == one) ? day02_part1(input) : day02_part2(input);
     free_array((void**)input);
     return r;
 }
