+++ 2015/src/day04/day04_test.c
@@ -2,6 +2,7 @@
 #include "testing.h"
 #include "utils.h"
 
+#include <stdint.h>
 #include <stdlib.h>
 
 const test_entry* day04_tests[] = {
@@ -12,9 +13,9 @@ const test_entry* day04_tests[] = {
     nullptr
 };
 
-int day04_test_run(const test_entry* te) {
-    char* input = file_to_string(te->file);
-    int   r     = (te->p == one) ? day04_part1(input) : day04_part2(input);
+int64_t day04_test_run(const test_entry* te) {
+    char*   input = file_to_string(te->file);
+    int64_t r     = (te->p == one) ? day04_part1(input) : day04_part2(input);
     free(input);
     return r;
 }
